requirements.txt
pytesseract
pdf2image
Pillow
opencv-python
python-dotenv
streamit